Create language.py to generate sender language
Contains a Language class to represent which kind of signals the sender will send to communicate location and orientation, and in which order. Function, ext_trajects() generates all (extensional) trajectories over the grid that a sender might send to communicate a goal to the receiver, given the sender's start and finish positions.